% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfixer.R
\name{ggfix}
\alias{ggfix}
\title{ggfix}
\usage{
ggfix(x)
}
\arguments{
\item{x}{The ggplot visualization you wish to fix}
}
\value{
A touched up visualization of visua \code{x}
}
\description{
ggfix
}
\examples{
library(ggplot2)
x<- ggplot(data = mtcars, aes(x = hp, y = mpg))+
geom_point()+ ggtitle("Mpg over Hp")
x
ggfix(x)
}
